Move input to separate file
